refactor: Split addon code into separate modules
Split addon handling code into separate modules:
- AddonBase.py - Base classes AddonBase and AddonError
- AddonBundle.py - Bundle class for handling addon packages
- AddonManifest.py - Manifest class and manifest handling

This improves code organization by:
- Separating concerns into logical modules
- Reducing complexity of __init__.py

No functional changes, only code reorganization.
