+++ source/addonStore/models/addon.py
@@ -22,6 +22,8 @@ from requests.structures import CaseInsensitiveDict
 
 import addonAPIVersion
 from NVDAState import WritePaths
+from addonHandler.AddonManifest import AddonManifest
+from addonHandler.addonBase import AddonBase as AddonHandlerBaseModel
 
 from .channel import Channel
 from .status import SupportsAddonState
@@ -33,8 +35,6 @@ from .version import (
 if TYPE_CHECKING:
 	from addonHandler import (  # noqa: F401
 		Addon as AddonHandlerModel,
-		AddonBase as AddonHandlerBaseModel,
-		AddonManifest,
 	)
 
 	AddonGUICollectionT = Dict[Channel, CaseInsensitiveDict["_AddonGUIModel"]]
